Split OHCI helpers and rename xHCI files
