+++ rom/usb/pciusb/ohci/ohcichip.c
@@ -99,7 +99,6 @@ static void ohciFreeTDChain(struct PCIController *hc, struct OhciTD *nextotd)
 static void ohciFreeEDContext(struct PCIController *hc, struct IOUsbHWReq *ioreq)
 {
     struct PCIDevice *base = hc->hc_Device;
-    struct PCIUnit *unit = hc->hc_Unit;
     struct OhciED *oed = ioreq->iouh_DriverPrivate1;
     UWORD devadrep;
     UWORD dir;
@@ -924,44 +923,6 @@ static void ohciScheduleIntTDs(struct PCIController *hc)
     }
 }
 
-static void ohciScheduleIsoTDs(struct PCIController *hc)
-{
-    struct PCIUnit *unit = hc->hc_Unit;
-    struct IOUsbHWReq *ioreq;
-    UWORD devadrep;
-
-    pciusbOHCIDebug("OHCI", "Scheduling new ISO transfers...\n");
-    ioreq = (struct IOUsbHWReq *)hc->hc_IsoXFerQueue.lh_Head;
-    while (((struct Node *)ioreq)->ln_Succ) {
-        devadrep = (ioreq->iouh_DevAddr << 5) + ioreq->iouh_Endpoint +
-                   ((ioreq->iouh_Dir == UHDIR_IN) ? 0x10 : 0);
-        if (unit->hu_DevBusyReq[devadrep]) {
-            ioreq = (struct IOUsbHWReq *)((struct Node *)ioreq)->ln_Succ;
-            continue;
-        }
-
-        struct RTIsoNode *rtn = pciusbAllocStdIsoNode(hc, ioreq);
-        if (!rtn)
-            break;
-
-        Remove(&ioreq->iouh_Req.io_Message.mn_Node);
-
-        if (ohciInitIsochIO(hc, rtn) != RC_OK) {
-            pciusbFreeStdIsoNode(hc, rtn);
-            ioreq->iouh_Req.io_Error = UHIOERR_HOSTERROR;
-            ReplyMsg(&ioreq->iouh_Req.io_Message);
-            ioreq = (struct IOUsbHWReq *)hc->hc_IsoXFerQueue.lh_Head;
-            continue;
-        }
-
-        ohciQueueIsochIO(hc, rtn);
-        AddTail((struct List *)&hc->hc_RTIsoHandlers, (struct Node *)&rtn->rtn_Node);
-        unit->hu_DevBusyReq[devadrep] = ioreq;
-        ohciStartIsochIO(hc, rtn);
-
-        ioreq = (struct IOUsbHWReq *)hc->hc_IsoXFerQueue.lh_Head;
-    }
-}
 
 static ULONG ohciScheduleBulkTDs(struct PCIController *hc)
 {
@@ -1260,51 +1221,7 @@ static AROS_INTH1(ohciIntCode, struct PCIController *, hc)
             pciusbOHCIDebug("OHCI", "HCI 0x%p: Frame Counter Rollover %ld\n", hc, hc->hc_FrameCounter);
         }
         if(intr & OISF_HUBCHANGE) {
-            UWORD hciport;
-            ULONG oldval;
-            UWORD portreg = OHCI_PORTSTATUS;
-            BOOL clearbits = FALSE;
-
-            if(READREG32_LE(hc->hc_RegBase, OHCI_INTSTATUS) & OISF_HUBCHANGE) {
-                // some OHCI implementations will keep the interrupt bit stuck until
-                // all port changes have been cleared, which is wrong according to the
-                // OHCI spec. As a workaround we will clear all change bits, which should
-                // be no problem as the port changes are reflected in the PortChangeMap
-                // array.
-                clearbits = TRUE;
-            }
-            for(hciport = 0; hciport < hc->hc_NumPorts; hciport++, portreg += 4) {
-                oldval = READREG32_LE(hc->hc_RegBase, portreg);
-                if(oldval & OHPF_OVERCURRENTCHG) {
-                    hc->hc_PortChangeMap[hciport] |= UPSF_PORT_OVER_CURRENT;
-                }
-                if(oldval & OHPF_RESETCHANGE) {
-                    hc->hc_PortChangeMap[hciport] |= UPSF_PORT_RESET;
-                }
-                if(oldval & OHPF_ENABLECHANGE) {
-                    hc->hc_PortChangeMap[hciport] |= UPSF_PORT_ENABLE;
-                }
-                if(oldval & OHPF_CONNECTCHANGE) {
-                    hc->hc_PortChangeMap[hciport] |= UPSF_PORT_CONNECTION;
-                }
-                if(oldval & OHPF_RESUMEDTX) {
-                    hc->hc_PortChangeMap[hciport] |= UPSF_PORT_SUSPEND;
-                }
-                if(clearbits) {
-                    WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_CONNECTCHANGE|OHPF_ENABLECHANGE|OHPF_RESUMEDTX|OHPF_OVERCURRENTCHG|OHPF_RESETCHANGE);
-                }
-
-                pciusbOHCIDebug("OHCI", "PCI Int Port %ld (glob %ld) Change %08lx\n", hciport, hc->hc_PortNum[hciport] + 1, oldval);
-                if(hc->hc_PortChangeMap[hciport]) {
-                    unit->hu_RootPortChanges |= 1UL<<(hc->hc_PortNum[hciport] + 1);
-                }
-            }
-            uhwCheckRootHubChanges(unit);
-            if(clearbits) {
-                // again try to get rid of any bits that may be causing the interrupt
-                WRITEREG32_LE(hc->hc_RegBase, OHCI_HUBSTATUS, OHSF_OVERCURRENTCHG);
-                WRITEREG32_LE(hc->hc_RegBase, OHCI_INTSTATUS, OISF_HUBCHANGE);
-            }
+            ohciCheckPortStatusChange(hc);
         }
         if(intr & OISF_DONEHEAD) {
             pciusbOHCIDebug("OHCI", "DoneHead Frame=%ld\n", READREG32_LE(hc->hc_RegBase, OHCI_FRAMECOUNT));
@@ -1423,313 +1340,10 @@ void ohciAbortRequest(struct PCIController *hc, struct IOUsbHWReq *ioreq)
     ohciEnableInt(hc, OISF_SOF);
 }
 
-WORD ohciInitIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    struct PTDNode *ptd0 = NULL;
-    struct PTDNode *ptd1 = NULL;
-    struct OhciIsoTD *oitd0 = NULL;
-    struct OhciIsoTD *oitd1 = NULL;
-    struct OhciED *oed = NULL;
-    UWORD idx;
-    UWORD ptdcount = rtn->rtn_PTDCount;
 
-    pciusbOHCIDebug("OHCI", "%s()\n", __func__);
 
-    if(!rtn->rtn_PTDs || ptdcount < 2)
-        return(UHIOERR_BADPARAMS);
 
-    for(idx = 0; idx < ptdcount; idx++)
-        rtn->rtn_PTDs[idx] = NULL;
 
-    ptd0 = AllocMem(sizeof(*ptd0), MEMF_CLEAR);
-    ptd1 = AllocMem(sizeof(*ptd1), MEMF_CLEAR);
-    if(!ptd0 || !ptd1) {
-        if(ptd0)
-            FreeMem(ptd0, sizeof(*ptd0));
-        if(ptd1)
-            FreeMem(ptd1, sizeof(*ptd1));
-        return(UHIOERR_OUTOFMEMORY);
-    }
-
-    oitd0 = ohciAllocIsoTD(hc);
-    oitd1 = ohciAllocIsoTD(hc);
-    oed = ohciAllocED(hc);
-    if(!oitd0 || !oitd1 || !oed) {
-        if(oitd0)
-            ohciFreeIsoTD(hc, oitd0);
-        if(oitd1)
-            ohciFreeIsoTD(hc, oitd1);
-        if(oed)
-            ohciFreeED(hc, oed);
-        FreeMem(ptd0, sizeof(*ptd0));
-        FreeMem(ptd1, sizeof(*ptd1));
-        return(UHIOERR_OUTOFMEMORY);
-    }
-
-    ptd0->ptd_Descriptor = oitd0;
-    ptd0->ptd_Phys = READMEM32_LE(&oitd0->oitd_Self);
-    ptd1->ptd_Descriptor = oitd1;
-    ptd1->ptd_Phys = READMEM32_LE(&oitd1->oitd_Self);
-
-    CONSTWRITEMEM32_LE(&oed->oed_EPCaps, OECF_SKIP|OECF_ISO);
-    WRITEMEM32_LE(&oed->oed_HeadPtr, 0);
-    WRITEMEM32_LE(&oed->oed_TailPtr, 0);
-    oed->oed_FirstTD = NULL;
-    oed->oed_IOReq = &rtn->rtn_IOReq;
-
-    rtn->rtn_IOReq.iouh_DriverPrivate1 = oed;
-    rtn->rtn_PTDs[0] = ptd0;
-    rtn->rtn_PTDs[1] = ptd1;
-    rtn->rtn_NextPTD = 0;
-    rtn->rtn_BounceBuffer = NULL;
-
-    return RC_OK;
-}
-
-WORD ohciQueueIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    struct PTDNode *ptd;
-    struct OhciIsoTD *oitd;
-    APTR dmabuffer;
-    ULONG phys;
-    ULONG frame;
-    ULONG ctrl;
-    UWORD pktcnt;
-    UWORD pktidx;
-    UWORD remaining;
-    UWORD offset;
-    struct IOUsbHWReq *ioreq = pciusbIsoGetIOReq(rtn);
-    struct IOUsbHWRTIso *urti = rtn->rtn_RTIso;
-    ULONG interval;
-
-    pciusbOHCIDebug("OHCI", "%s()\n", __func__);
-
-    if (urti) {
-        if (ioreq->iouh_Dir == UHDIR_IN) {
-            if (urti->urti_InReqHook)
-                CallHookPkt(urti->urti_InReqHook, rtn, &rtn->rtn_BufferReq);
-        } else {
-            if (urti->urti_OutReqHook)
-                CallHookPkt(urti->urti_OutReqHook, rtn, &rtn->rtn_BufferReq);
-        }
-    } else {
-        rtn->rtn_BufferReq.ubr_Buffer = ioreq->iouh_Data;
-        rtn->rtn_BufferReq.ubr_Length = ioreq->iouh_Length;
-        rtn->rtn_BufferReq.ubr_Frame = ioreq->iouh_Frame;
-        rtn->rtn_BufferReq.ubr_Flags = 0;
-    }
-
-    if (!rtn->rtn_BufferReq.ubr_Length)
-        rtn->rtn_BufferReq.ubr_Length = ioreq->iouh_Length;
-
-    interval = ioreq->iouh_Interval ? ioreq->iouh_Interval : 1;
-    if (!rtn->rtn_BufferReq.ubr_Frame) {
-        ULONG framecnt = READREG32_LE(hc->hc_RegBase, OHCI_FRAMECOUNT) & 0xffff;
-        ULONG next = rtn->rtn_NextFrame ? rtn->rtn_NextFrame : ((framecnt + interval) & 0xffff);
-        rtn->rtn_BufferReq.ubr_Frame = next;
-    }
-    rtn->rtn_NextFrame = (rtn->rtn_BufferReq.ubr_Frame + interval) & 0xffff;
-
-    dmabuffer = rtn->rtn_BufferReq.ubr_Buffer;
-#if __WORDSIZE == 64
-    if(!dmabuffer)
-        return RC_OK;
-
-    dmabuffer = usbGetBuffer(dmabuffer, rtn->rtn_BufferReq.ubr_Length,
-                             rtn->rtn_IOReq.iouh_Dir);
-    if(!dmabuffer)
-        return UHIOERR_OUTOFMEMORY;
-
-    if(dmabuffer != rtn->rtn_BufferReq.ubr_Buffer &&
-            rtn->rtn_IOReq.iouh_Dir == UHDIR_OUT) {
-        CopyMem(rtn->rtn_BufferReq.ubr_Buffer, dmabuffer, rtn->rtn_BufferReq.ubr_Length);
-    }
-#endif
-    rtn->rtn_BounceBuffer = (dmabuffer != rtn->rtn_BufferReq.ubr_Buffer) ? dmabuffer : NULL;
-
-    ptd = rtn->rtn_PTDs[rtn->rtn_NextPTD];
-    if(!ptd)
-        return(UHIOERR_BADPARAMS);
-
-    oitd = (struct OhciIsoTD *)ptd->ptd_Descriptor;
-    if(!oitd)
-        return(UHIOERR_BADPARAMS);
-
-    if(!dmabuffer || !rtn->rtn_BufferReq.ubr_Length)
-        return RC_OK;
-
-    phys = (ULONG)(IPTR)pciGetPhysical(hc, dmabuffer);
-    frame = rtn->rtn_BufferReq.ubr_Frame;
-
-    pktcnt = (rtn->rtn_BufferReq.ubr_Length + rtn->rtn_IOReq.iouh_MaxPktSize - 1) /
-             rtn->rtn_IOReq.iouh_MaxPktSize;
-    if(pktcnt > 8)
-        pktcnt = 8;
-
-    ctrl = (frame << OITCS_STARTINGFRAME) | OITF_NOINT |
-           ((pktcnt - 1) << OITCS_FRAMECOUNT) | OITF_CC_NOTACCESSED;
-
-    oitd->oitd_Succ = NULL;
-    oitd->oitd_ED = (struct OhciED *)rtn->rtn_IOReq.iouh_DriverPrivate1;
-    oitd->oitd_Length = rtn->rtn_BufferReq.ubr_Length;
-    WRITEMEM32_LE(&oitd->oitd_Ctrl, ctrl);
-    WRITEMEM32_LE(&oitd->oitd_BufferPage0, phys & ~0xfff);
-    WRITEMEM32_LE(&oitd->oitd_NextTD, 0);
-
-    remaining = rtn->rtn_BufferReq.ubr_Length;
-    offset = phys & 0xfff;
-    for(pktidx = 0; pktidx < pktcnt; pktidx++) {
-        UWORD pktlen = remaining;
-        if(pktlen > rtn->rtn_IOReq.iouh_MaxPktSize)
-            pktlen = rtn->rtn_IOReq.iouh_MaxPktSize;
-
-        oitd->oitd_Offset[pktidx] = offset | OITM_PSW_CC;
-        offset += pktlen;
-        remaining -= pktlen;
-    }
-    while(pktidx < 8) {
-        oitd->oitd_Offset[pktidx++] = OITM_PSW_CC;
-    }
-
-    WRITEMEM32_LE(&oitd->oitd_BufferEnd, phys + offset - 1);
-
-    CacheClearE(oitd, sizeof(*oitd), CACRF_ClearD);
-
-    ptd->ptd_Length = offset;
-    ptd->ptd_FrameIdx = frame;
-    ptd->ptd_Flags = PTDF_BUFFER_VALID;
-
-    rtn->rtn_NextPTD ^= 1;
-
-    return RC_OK;
-}
-
-void ohciStartIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    struct OhciHCPrivate *ohcihcp = (struct OhciHCPrivate *)hc->hc_CPrivate;
-    struct IOUsbHWReq *ioreq = pciusbIsoGetIOReq(rtn);
-    struct OhciED *oed = (struct OhciED *)rtn->rtn_IOReq.iouh_DriverPrivate1;
-    UWORD idx;
-    struct OhciED *intoed;
-
-    pciusbOHCIDebug("OHCI", "%s()\n", __func__);
-
-    if(!oed)
-        return;
-
-    WRITEMEM32_LE(&oed->oed_EPCaps,
-                  OECF_ISO |
-                  ((rtn->rtn_IOReq.iouh_DevAddr << OECS_DEVADDR) & OECM_DEVADDR) |
-                  ((rtn->rtn_IOReq.iouh_Endpoint << OECS_ENDPOINT) & OECM_ENDPOINT) |
-                  (rtn->rtn_IOReq.iouh_MaxPktSize << OECS_MAXPKTLEN));
-    oed->oed_IOReq = ioreq;
-    ioreq->iouh_DriverPrivate2 = rtn;
-
-    for(idx = 0; idx < 2; idx++) {
-        struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-        struct OhciIsoTD *oitd;
-
-        if(!ptd || !(ptd->ptd_Flags & PTDF_BUFFER_VALID))
-            continue;
-
-        oitd = (struct OhciIsoTD *)ptd->ptd_Descriptor;
-        if(!oitd)
-            continue;
-
-        WRITEMEM32_LE(&oitd->oitd_NextTD, ohcihcp->ohc_OhciTermTD->otd_Self);
-        oitd->oitd_ED = oed;
-        CacheClearE(oitd, sizeof(*oitd), CACRF_ClearD);
-
-        oed->oed_FirstTD = (struct OhciTD *)oitd;
-        WRITEMEM32_LE(&oed->oed_TailPtr, ohcihcp->ohc_OhciTermTD->otd_Self);
-        WRITEMEM32_LE(&oed->oed_HeadPtr, READMEM32_LE(&oitd->oitd_Self));
-        CacheClearE(&oed->oed_EPCaps, 16, CACRF_ClearD);
-
-        ptd->ptd_Flags |= PTDF_ACTIVE;
-    }
-
-    if(!oed->oed_Pred) {
-        if(rtn->rtn_IOReq.iouh_Interval >= 31) {
-            intoed = ohcihcp->ohc_OhciIntED[4];
-        } else {
-            UWORD cnt = 0;
-            do {
-                intoed = ohcihcp->ohc_OhciIntED[cnt++];
-            } while(rtn->rtn_IOReq.iouh_Interval >= (1 << cnt));
-        }
-
-        Disable();
-        oed->oed_Succ = intoed->oed_Succ;
-        oed->oed_Pred = intoed;
-        oed->oed_NextED = intoed->oed_Succ->oed_Self;
-        intoed->oed_Succ->oed_Pred = oed;
-        intoed->oed_Succ = oed;
-        intoed->oed_NextED = oed->oed_Self;
-        CacheClearE(&oed->oed_EPCaps, 16, CACRF_ClearD);
-        CacheClearE(&intoed->oed_EPCaps, 16, CACRF_ClearD);
-        ohciUpdateIntTree(hc);
-        Enable();
-    }
-
-    WRITEMEM32_LE(&oed->oed_EPCaps, READMEM32_LE(&oed->oed_EPCaps) & ~OECF_SKIP);
-    CacheClearE(&oed->oed_EPCaps, 16, CACRF_ClearD);
-}
-
-void ohciStopIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    struct OhciED *oed = (struct OhciED *)rtn->rtn_IOReq.iouh_DriverPrivate1;
-    UWORD ptdcount = rtn->rtn_PTDCount;
-    UWORD limit = (ptdcount < 2) ? ptdcount : 2;
-    UWORD idx;
-
-    pciusbOHCIDebug("OHCI", "%s()\n", __func__);
-
-    if(oed && !(READMEM32_LE(&oed->oed_EPCaps) & OECF_SKIP)) {
-        ohciDisableED(oed);
-        ohciEnableInt(hc, OISF_SOF);
-    }
-
-    for(idx = 0; idx < limit; idx++) {
-        if(rtn->rtn_PTDs[idx])
-            rtn->rtn_PTDs[idx]->ptd_Flags &= ~(PTDF_ACTIVE|PTDF_BUFFER_VALID);
-    }
-
-    if(rtn->rtn_BounceBuffer &&
-            rtn->rtn_BounceBuffer != rtn->rtn_BufferReq.ubr_Buffer) {
-        usbReleaseBuffer(rtn->rtn_BounceBuffer, rtn->rtn_BufferReq.ubr_Buffer,
-                         rtn->rtn_BufferReq.ubr_Length, rtn->rtn_IOReq.iouh_Dir);
-        rtn->rtn_BounceBuffer = NULL;
-    }
-}
-
-void ohciFreeIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    struct OhciED *oed = (struct OhciED *)rtn->rtn_IOReq.iouh_DriverPrivate1;
-    UWORD idx;
-    UWORD ptdcount = rtn->rtn_PTDCount;
-    UWORD limit = (ptdcount < 2) ? ptdcount : 2;
-
-    pciusbOHCIDebug("OHCI", "%s()\n", __func__);
-
-    ohciStopIsochIO(hc, rtn);
-
-    if(oed) {
-        ohciFreeED(hc, oed);
-        rtn->rtn_IOReq.iouh_DriverPrivate1 = NULL;
-    }
-
-    for(idx = 0; idx < limit; idx++) {
-        struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-        if(ptd) {
-            if(ptd->ptd_Descriptor)
-                ohciFreeIsoTD(hc, (struct OhciIsoTD *)ptd->ptd_Descriptor);
-            FreeMem(ptd, sizeof(*ptd));
-            rtn->rtn_PTDs[idx] = NULL;
-        }
-    }
-
-    rtn->rtn_BounceBuffer = NULL;
-}
 
 BOOL ohciInit(struct PCIController *hc, struct PCIUnit *hu)
 {
@@ -2090,175 +1704,4 @@ void ohciFree(struct PCIController *hc, struct PCIUnit *hu)
     KPRINTF(20, "Shutting down OHCI done.\n");
 }
 
-/* ** Root hub support functions ** */
 
-BOOL ohciSetFeature(struct PCIUnit *unit, struct PCIController *hc, UWORD hciport, UWORD idx, UWORD val, WORD *retval)
-{
-    UWORD portreg = OHCI_PORTSTATUS + (hciport<<2);
-    ULONG oldval = READREG32_LE(hc->hc_RegBase, portreg);
-    BOOL cmdgood = FALSE;
-
-    pciusbOHCIDebug("OHCI", "%s(0x%p, 0x%p, %04x, %04x, %04x, 0x%p)\n", __func__, unit, hc, hciport, idx, val, retval);
-
-    switch(val) {
-    /* case UFS_PORT_CONNECTION: not possible */
-    case UFS_PORT_ENABLE:
-        pciusbOHCIDebug("OHCI", "Enabling Port (%s)\n", oldval & OHPF_PORTENABLE ? "already" : "ok");
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTENABLE);
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_SUSPEND:
-        pciusbOHCIDebug("OHCI", "Suspending Port (%s)\n", oldval & OHPF_PORTSUSPEND ? "already" : "ok");
-        //hc->hc_PortChangeMap[hciport] |= UPSF_PORT_SUSPEND; // manually fake suspend change
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTSUSPEND);
-        cmdgood = TRUE;
-        break;
-
-    /* case UFS_PORT_OVER_CURRENT: not possible */
-    case UFS_PORT_RESET:
-        pciusbOHCIDebug("OHCI", "Resetting Port (%s)\n", oldval & OHPF_PORTRESET ? "already" : "ok");
-        // make sure we have at least 50ms of reset time here, as required for a root hub port
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTRESET);
-        uhwDelayMS(10, unit->hu_TimerReq);
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTRESET);
-        uhwDelayMS(10, unit->hu_TimerReq);
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTRESET);
-        uhwDelayMS(10, unit->hu_TimerReq);
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTRESET);
-        uhwDelayMS(10, unit->hu_TimerReq);
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTRESET);
-        uhwDelayMS(15, unit->hu_TimerReq);
-        oldval = READREG32_LE(hc->hc_RegBase, portreg);
-        pciusbOHCIDebug("OHCI", "Reset release (%s %s)\n", oldval & OHPF_PORTRESET ? "didn't turn off" : "okay",
-                    oldval & OHPF_PORTENABLE ? "enabled" : "not enabled");
-        if(oldval & OHPF_PORTRESET) {
-            uhwDelayMS(40, unit->hu_TimerReq);
-            oldval = READREG32_LE(hc->hc_RegBase, portreg);
-            pciusbOHCIDebug("OHCI", "Reset 2nd release (%s %s)\n", oldval & OHPF_PORTRESET ? "didn't turn off" : "okay",
-                        oldval & OHPF_PORTENABLE ? "enabled" : "still not enabled");
-        }
-        // make enumeration possible
-        unit->hu_DevControllers[0] = hc;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_POWER:
-        pciusbOHCIDebug("OHCI", "Powering Port (%s)\n", oldval & OHPF_PORTPOWER ? "already" : "ok");
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTPOWER);
-        cmdgood = TRUE;
-        break;
-
-        /* case UFS_PORT_LOW_SPEED: not possible */
-        /* case UFS_C_PORT_CONNECTION:
-        case UFS_C_PORT_ENABLE:
-        case UFS_C_PORT_SUSPEND:
-        case UFS_C_PORT_OVER_CURRENT:
-        case UFS_C_PORT_RESET: */
-    }
-    return cmdgood;
-}
-
-BOOL ohciClearFeature(struct PCIUnit *unit, struct PCIController *hc, UWORD hciport, UWORD idx, UWORD val, WORD *retval)
-{
-    UWORD portreg = OHCI_PORTSTATUS + (hciport<<2);
-    ULONG __unused oldval = READREG32_LE(hc->hc_RegBase, portreg);
-    BOOL cmdgood = FALSE;
-
-    pciusbOHCIDebug("OHCI", "%s(0x%p, 0x%p, %04x, %04x, %04x, 0x%p)\n", __func__, unit, hc, hciport, idx, val, retval);
-
-    switch(val) {
-    case UFS_PORT_ENABLE:
-        pciusbOHCIDebug("OHCI", "Disabling Port (%s)\n", oldval & OHPF_PORTENABLE ? "ok" : "already");
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTDISABLE);
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_SUSPEND:
-        pciusbOHCIDebug("OHCI", "Resuming Port (%s)\n", oldval & OHPF_PORTSUSPEND ? "ok" : "already");
-        //hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_SUSPEND; // manually fake suspend change
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_RESUME);
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_POWER:
-        pciusbOHCIDebug("OHCI", "Unpowering Port (%s)\n", oldval & OHPF_PORTPOWER ? "ok" : "already");
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_PORTUNPOWER);
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_CONNECTION:
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_CONNECTCHANGE);
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_CONNECTION;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_ENABLE:
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_ENABLECHANGE);
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_ENABLE;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_SUSPEND:
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_RESUMEDTX);
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_SUSPEND;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_OVER_CURRENT:
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_OVERCURRENTCHG);
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_OVER_CURRENT;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_RESET:
-        WRITEREG32_LE(hc->hc_RegBase, portreg, OHPF_RESETCHANGE);
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_RESET;
-        cmdgood = TRUE;
-        break;
-    }
-    return cmdgood;
-}
-
-BOOL ohciGetStatus(struct PCIController *hc, UWORD *mptr, UWORD hciport, UWORD idx, WORD *retval)
-{
-    UWORD portreg = OHCI_PORTSTATUS + (hciport<<2);
-    ULONG oldval = READREG32_LE(hc->hc_RegBase, portreg);
-
-    pciusbOHCIDebug("OHCI", "%s(0x%p, 0x%p, %04x, %04x, 0x%p)\n", __func__, hc, mptr, hciport, idx, retval);
-
-    *mptr = 0;
-    if(oldval & OHPF_PORTPOWER) *mptr |= AROS_WORD2LE(UPSF_PORT_POWER);
-    if(oldval & OHPF_OVERCURRENT) *mptr |= AROS_WORD2LE(UPSF_PORT_OVER_CURRENT);
-    if(oldval & OHPF_PORTCONNECTED) *mptr |= AROS_WORD2LE(UPSF_PORT_CONNECTION);
-    if(oldval & OHPF_PORTENABLE) *mptr |= AROS_WORD2LE(UPSF_PORT_ENABLE);
-    if(oldval & OHPF_LOWSPEED) *mptr |= AROS_WORD2LE(UPSF_PORT_LOW_SPEED);
-    if(oldval & OHPF_PORTRESET) *mptr |= AROS_WORD2LE(UPSF_PORT_RESET);
-    if(oldval & OHPF_PORTSUSPEND) *mptr |= AROS_WORD2LE(UPSF_PORT_SUSPEND);
-
-    pciusbOHCIDebug("OHCI", "Port %ld (glob. %ld) is %s\n", hciport, idx, oldval & OHPF_LOWSPEED ? "LOWSPEED" : "FULLSPEED");
-    pciusbOHCIDebug("OHCI", "Port %ld Status %08lx (%08lx)\n", idx, *mptr, oldval);
-
-    mptr++;
-    if(oldval & OHPF_OVERCURRENTCHG) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_OVER_CURRENT;
-    }
-    if(oldval & OHPF_RESETCHANGE) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_RESET;
-    }
-    if(oldval & OHPF_ENABLECHANGE) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_ENABLE;
-    }
-    if(oldval & OHPF_CONNECTCHANGE) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_CONNECTION;
-    }
-    if(oldval & OHPF_RESUMEDTX) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_SUSPEND;
-    }
-
-    *mptr = AROS_WORD2LE(hc->hc_PortChangeMap[hciport]);
-
-    pciusbOHCIDebug("OHCI", "Port %ld Change %08lx\n", idx, *mptr);
-
-    return TRUE;
-}
